Disable cake2 and refactor tests
- Disable cake2 by default
- Refactor tests to account for flag
- Fix problem with component completion being disabled
- Add todo for nested component completion
