+++ src/test/kotlin/com/daveme/chocolateCakePHP/test/cake3/ControllerLineMarkerTest.kt
@@ -1,16 +1,17 @@
-package com.daveme.chocolateCakePHP.test
+package com.daveme.chocolateCakePHP.test.cake3
 
 import com.daveme.chocolateCakePHP.Settings
 import com.daveme.chocolateCakePHP.controller.ControllerMethodLineMarker
+import com.daveme.chocolateCakePHP.test.BaseTestCase
 import com.intellij.icons.AllIcons
 import com.intellij.psi.util.PsiTreeUtil
 import org.junit.Test
 import com.jetbrains.php.lang.psi.elements.Method
 import com.jetbrains.php.lang.psi.elements.MethodReference
 
-class ControllerLineMarkerTest : BaseTestCase() {
+class ControllerLineMarkerTest : Cake3BaseTestCase() {
 
-    private fun prepareTest() {
+    override fun prepareTest() {
         // change app directory:
         val originalSettings = Settings.getInstance(myFixture.project)
         val newState = originalSettings.state.copy()
@@ -20,8 +21,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that line markers contain one entry for each type of view`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
@@ -66,8 +65,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that line marker navigates to explicit render() calls`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
@@ -115,8 +112,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that line marker navigates to explicit render() calls when nested`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
@@ -165,8 +160,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that line marker adds markers to render calls`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
@@ -213,8 +206,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that a line marker is added to the method name when a corresponding view file doesn't exist`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
@@ -251,8 +242,6 @@ class ControllerLineMarkerTest : BaseTestCase() {
 
     @Test
     fun `test that a line marker is added next to render call when a corresponding view file doesn't exist`() {
-        prepareTest();
-
         val files = myFixture.configureByFiles(
             "cake3/src/Controller/AppController.php",
             "cake3/vendor/cakephp.php",
